+++ test/integration/api_query_test.lua
@@ -5,12 +5,12 @@ local g = t.group()
 local helpers = require('test.helper')
 local log = require('log')
 
-g.before_all = function()
+g.before_all(function()
     g.cluster = helpers.Cluster:new({
         datadir = fio.tempdir(),
         server_command = helpers.entrypoint('srv_basic'),
         use_vshard = true,
-        cookie = require('digest').urandom(6):hex(),
+        cookie = helpers.random_cookie(),
 
         replicasets = {
             {
@@ -117,7 +117,7 @@ g.before_all = function()
             variables = {uri = g.server.advertise_uri},
         })
     end)
-end
+end)
 
 g.before_each(function()
     helpers.retrying({}, function()
@@ -125,12 +125,12 @@ g.before_each(function()
     end)
 end)
 
-g.after_all = function()
+g.after_all(function()
     g.cluster:stop()
     g.server:stop()
     fio.rmtree(g.cluster.datadir)
     fio.rmtree(g.server.workdir)
-end
+end)
 
 local function fields_from_map(map, field_key)
     local data_arr = {}
@@ -768,4 +768,50 @@ function g.test_stop_roles_on_shutdown()
         fio.path.exists(last_will_path),
         true
     )
+end
+
+function g.test_cartridge_get_topology_iproto()
+    local res = g.cluster:server('router'):exec(function()
+        require('membership').probe_uri('localhost:13303')
+        return require('cartridge.lua-api.get-topology').get_servers()
+    end)
+
+    local expected = {{
+            alias = "router",
+            disabled = false,
+            labels = {},
+            priority = 1,
+            status = "healthy",
+            replicaset_uuid = "aaaaaaaa-0000-0000-0000-000000000000",
+            uri = "localhost:13301",
+            uuid = "aaaaaaaa-aaaa-0000-0000-000000000001",
+        }, {
+            alias = "storage",
+            disabled = false,
+            labels = {},
+            priority = 1,
+            replicaset_uuid = "bbbbbbbb-0000-0000-0000-000000000000",
+            uri = "localhost:13302",
+            uuid = "bbbbbbbb-bbbb-0000-0000-000000000001",
+        }, {
+            alias = "spare",
+            uri = "localhost:13303",
+            uuid = "",
+        }, {
+            alias = "storage-2",
+            disabled = false,
+            labels = {},
+            priority = 2,
+            replicaset_uuid = "bbbbbbbb-0000-0000-0000-000000000000",
+            uri = "localhost:13304",
+            uuid = "bbbbbbbb-bbbb-0000-0000-000000000002",
+    }}
+
+    t.assert_equals(#expected, #res)
+
+    table.sort(expected, function(a,b) return a.uri < b.uri end)
+    table.sort(res, function(a,b) return a.uri < b.uri end)
+    for i, exp_server in ipairs(expected) do
+        t.assert_covers(res[i], exp_server)
+    end
 end
